function/requirements.txt
azure-functions
azure-storage-blob>=12.0.0
qrcode[pil]
Pillow
